fix: align MealLibraryItem model, schema, and API with actual fields
- Change Calories from int to float64 to match schema DOUBLE PRECISION
- Add category and ingredients columns to meal_library_items schema
- Update repository SQL (CRUD + batch replace) and scanItem accordingly
- Add category and ingredients to OpenAPI spec and generated types
- Update mappers to remove lossy int/float64 casts and wire new fields

+++ backend/internal/library/model.go
@@ -22,7 +22,7 @@ type MealLibraryItem struct {
 	ID          string    `json:"id" db:"id"`
 	UserID      string    `json:"user_id" db:"user_id"`
 	Name        string    `json:"name" db:"name"`
-	Calories    int       `json:"calories" db:"calories"`
+	Calories    float64   `json:"calories" db:"calories"`
 	Weight      float64   `json:"weight" db:"weight"`
 	Category    string    `json:"category" db:"category"`
 	Ingredients []string  `json:"ingredients" db:"ingredients"`
